コンポーネントの整理
