+++ components/utils.ts
@@ -1,0 +1,1 @@
+export const RESAS_BASE_URL = "https://opendata.resas-portal.go.jp/api/v1";
