+++ pages/index.tsx
@@ -1,15 +1,12 @@
 import Head from "next/head";
-import { useEffect, useState } from "react";
-import axios from "axios";
+import { useState } from "react";
 
 import styles from "../styles/Home.module.css";
 import { Prefecture } from "../models/Prefecture";
 import CustomLineChart from "../components/CustomLineChart";
-
-const BASE_URL = "https://opendata.resas-portal.go.jp/api/v1";
+import PrefectureSelector from "../components/PrefectureSelector";
 
 export default function Home() {
-  const [prefs, setPrefs] = useState<Prefecture[]>([]);
   const [selectedPrefs, setSelectedPrefs] = useState<Prefecture[]>([]);
 
   const handleChange = (value: Prefecture) => {
@@ -19,19 +16,6 @@ export default function Home() {
       setSelectedPrefs([...selectedPrefs, value]);
     }
   };
-
-  useEffect(() => {
-    const fetchPrefs = async () => {
-      const response = await axios.get(BASE_URL + "/prefectures", {
-        headers: {
-          "X-API-KEY": process.env.NEXT_PUBLIC_API_KEY,
-        },
-      });
-      console.log(response.data);
-      setPrefs(response.data.result as Prefecture[]);
-    };
-    fetchPrefs();
-  }, []);
 
   return (
     <div className={styles.container}>
@@ -46,18 +30,9 @@ export default function Home() {
 
       <main className={styles.main}>
         <h1 className={styles.title}>Population Chart</h1>
-        <div className={styles.checkbox_group}>
-          <ul>
-            {prefs.map((pref) => (
-              <li key={pref.prefCode}>
-                <label>
-                  <input type="checkbox" onChange={() => handleChange(pref)} />
-                  {pref.prefName}
-                </label>
-              </li>
-            ))}
-          </ul>
-        </div>
+        <>
+          <PrefectureSelector handleChange={handleChange} />
+        </>
         <div className={styles.chart}>
           <CustomLineChart selected={selectedPrefs} />
         </div>
