+++ styles/Home.module.css
@@ -38,16 +38,3 @@
   margin-top: 5rem;
   width: 100%;
 }
-
-.checkbox_group {
-  margin-top: 2rem;
-  column-count: 5;
-}
-
-.checkbox_group li {
-  list-style: none;
-}
-
-.checkbox_group input {
-  margin-right: 0.5rem;
-}
